requirements.txt
annotated-types==0.7.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
defusedxml==0.7.1
Flask==3.1.0
future==1.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
oic==1.7.0
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyjwkest==1.4.2
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3